split out existing callback_context tests to their own file
